api/pyproject.toml
[tool.poetry]
name = "plataforma_atus_publicao_express"
version = "0.1.0"
description = ""
authors = ["Felipe Ferreira <f3cdde@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
python-decouple = "^3.4"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
PyPDF2 = "^1.26.0"
pdf2image = "^1.16.0"
pdfCropMargins = "^1.0.5"
appdirs = "1.4.4"
attrs = "21.2.0"
autopep8 = "1.5.7"
CacheControl = "0.12.6"
cachy = "0.3.0"
certifi = "2020.12.5"
cffi = "1.14.5"
chardet = "4.0.0"
cleo = "0.8.1"
clikit = "0.6.2"
crashtest = "0.3.1"
cryptography = "3.4.7"
distlib = "0.3.1"
dj-database-url = "0.5.0"
file-magic = "0.4.0"
filelock = "3.0.12"
flake8 = "3.9.2"
html5lib = "1.1"
idna = "2.10"
iniconfig = "1.1.1"
jeepney = "0.6.0"
keyring = "21.8.0"
lockfile = "0.12.2"
mccabe = "0.6.1"
msgpack = "1.0.2"
mypy = "0.812"
mypy-extensions = "0.4.3"
packaging = "20.9"
pastel = "0.2.1"
pexpect = "4.8.0"
pkginfo = "1.7.0"
pluggy = "0.13.1"
poetry = "1.1.6"
poetry-core = "1.0.3"
ptyprocess = "0.7.0"
py = "1.10.0"
pycodestyle = "2.7.0"
pycparser = "2.20"
pyflakes = "2.3.1"
pylev = "1.3.0"
pyparsing = "2.4.7"
pytest = "6.2.4"
requests = "2.25.1"
requests-toolbelt = "0.9.1"
SecretStorage = "3.3.1"
shellingham = "1.4.0"
six = "1.16.0"
sqlparse = "0.4.1"
toml = "0.10.2"
tomlkit = "0.7.2"
typed-ast = "1.4.3"
typing-extensions = "3.10.0.0"
urllib3 = "1.26.4"
virtualenv = "20.4.7"
webencodings = "0.5.1"
python-docx = { git = "https://github.com/phgrigorio/python-docx", branch = "master" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
